==> code/processing/02_processing_implicit_native_scores.Rmd <==
---
title: "Data processing"
author: "Jamie Cummins & Ian Hussey"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

```

```{r}

library(tidyverse)
library(parallel)
library(boot)

```

# Process IAT D score data

## Load data with tidying and exclusions already done (in the main processing script)

```{r}

all_exclusions_applied_iat_df <- read_rds("../../data/processed/data_trials_for_bootstrapping.rds") |>
  filter(measure %in% c("iat", "biat", "siat"))

```

# bootstrap 95% CIs and compare against neutral point 

```{r}

n_boots = 2000
subset_sample = FALSE # to reduce processing time for testing

if(TRUE){
  
  D_score <- function(data, i) {
    data_with_indexes <- data[i,] # boot function requires data and index
    b  <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "inconsistent"])
    a  <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "consistent"])
    mean_b <- mean(b)
    mean_a <- mean(a)
    sd <- sd(c(a, b))
    D <- (mean_b-mean_a)/sd
    return(D)
  }
  
  bootstrap_D_score <- function(data){
    
    require(dplyr)
    require(boot)
    
    fit <- 
      boot::boot(data      = data, 
                 statistic = D_score, 
                 R         = n_boots, 
                 sim       = "ordinary", 
                 stype     = "i",
                 parallel  = "multicore", 
                 ncpus     = parallel::detectCores()-1)
    
    # results <- boot::boot.ci(fit, conf = 0.95, type = c("bca"))
    # 
    # output <- 
    #   tibble(method   = "bca",
    #          estimate = fit$t0,
    #          ci_lower = results$bca[4],
    #          ci_upper = results$bca[5])
    # returns error: estimated adjustment 'w' is infinite
    
    results <- boot::boot.ci(fit, conf = 0.95, type = c("basic"))
    
    output <- 
      tibble(method   = "basic",
             estimate = fit$t0,
             ci_lower = results$basic[4],
             ci_upper = results$basic[5])
    
    return(output)
  }
  
  
  if(subset_sample == TRUE) {
    trials_to_process <- all_exclusions_applied_iat_df |>
      slice_head(n = 500000)
    } else {
    trials_to_process <- all_exclusions_applied_iat_df
  }
  
  # bootstrap D scores 
  D_scores <- trials_to_process |>
    mutate(session_id = as.factor(SESSION_ID),
           trial_consistency = as.character(trial_consistency)) |>
    
    # ensure a sufficient number of observations per person per measure per trial type
    group_by(session_id, measure, domain, trial_consistency) |>
    filter(n() >= 10) |>
    
    # remove those who do not have sufficient observations for both trials per measure
    group_by(session_id, domain, measure) |>
    filter(n_distinct(trial_consistency) == 2) |>
    
    # do bootstrapping scores
    group_by(session_id, measure, domain) |>
    do(bootstrap_D_score(data = .)) |>
    ungroup() |>
    mutate(sig = ifelse((ci_lower < 0 & ci_upper < 0) | (ci_lower > 0 & ci_upper > 0), TRUE, FALSE),
           ci_width = ci_upper - ci_lower) 
  
}

data_processed <- data_processed |>
  mutate(sig = ifelse((ci_lower < 0 & ci_upper < 0) | (ci_lower > 0 & ci_upper > 0), TRUE, FALSE),
         ci_width = ci_upper - ci_lower) 

```

# Process AMP scoring data

```{r}

all_exclusions_applied_amp_df <- read_rds("../../data/processed/data_trials_for_bootstrapping.rds") |>
  filter(measure %in% c("amp"))

```

## boostrap AMP scores

```{r}

n_boots <- 2000
subset_sample <- FALSE # optional for testing

if (TRUE) {
  
  # Function to compute AMP score (difference in means)
  AMP_score <- function(data, i) {
    data_with_indexes <- data[i, ] # Required format for boot
    b <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "inconsistent"])
    a <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "consistent"])
    mean_a <- mean(a)
    mean_b <- mean(b)
    amp <- mean_a - mean_b
    return(amp)
  }

  bootstrap_AMP_score <- function(data) {
    require(dplyr)
    require(boot)
    
    fit <- boot::boot(
      data      = data,
      statistic = AMP_score,
      R         = n_boots,
      sim       = "ordinary",
      stype     = "i",
      parallel  = "multicore",
      ncpus     = parallel::detectCores() - 1
    )

    results <- boot::boot.ci(fit, conf = 0.95, type = "basic")
    
    tibble(
      method   = "basic",
      estimate = fit$t0,
      ci_lower = results$basic[4],
      ci_upper = results$basic[5]
    )
  }

  # Subset for speed (optional)
  if (subset_sample == TRUE) {
    trials_to_process <- all_exclusions_applied_amp_df |>  
      slice_head(n = 500000)
  } else {
    trials_to_process <- all_exclusions_applied_amp_df
  }

  # Bootstrap AMP scores per participant + domain
  AMP_scores <- trials_to_process |>
    mutate(session_id = as.factor(SESSION_ID),
           trial_consistency = as.character(trial_consistency)) |>
    group_by(session_id, domain) |>
    filter(n_distinct(trial_consistency) == 2) |>  # Must have both conditions
    
    group_by(session_id, domain) |>
    do(bootstrap_AMP_score(data = .)) |>
    ungroup() |>
    mutate(
      sig = ifelse((ci_lower < 0 & ci_upper < 0) | (ci_lower > 0 & ci_upper > 0), TRUE, FALSE),
      ci_width = ci_upper - ci_lower
    )
}


```

# Process EPT scoring data

```{r}

all_exclusions_applied_ept_df <- read_rds("../../data/processed/data_trials_for_bootstrapping.rds") |>
  filter(measure %in% c("ept"))


```

### boostrap EPT score

```{r}

n_boots <- 2000
subset_sample <- FALSE  # for speed while testing

# Function to compute EPT score (RT_incongruent - RT_congruent)
EPT_score <- function(data, i) {
  data_with_indexes <- data[i, ]
  
  # Filter out invalid RTs and incorrect responses
  cleaned <- data_with_indexes |>
    filter(TRIAL_ERROR == 0, score >= 300, score <= 2000) 
  
  if (!all(c("consistent", "inconsistent") %in% cleaned$trial_consistency)) return(NA)
  
  mean_congruent <- mean(cleaned$score[cleaned$trial_consistency == "consistent"])
  mean_incongruent <- mean(cleaned$score[cleaned$trial_consistency == "inconsistent"])
  
  return(mean_incongruent - mean_congruent)
}

# Bootstrap wrapper
bootstrap_EPT_score <- function(data) {
  require(boot)
  require(dplyr)
  
  fit <- boot::boot(
    data = data,
    statistic = EPT_score,
    R = n_boots,
    sim = "ordinary",
    stype = "i",
    parallel = "multicore",
    ncpus = parallel::detectCores() - 1
  )
  
  results <- boot::boot.ci(fit, conf = 0.95, type = "basic")
  
  tibble(
    method   = "basic",
    estimate = fit$t0,
    ci_lower = results$basic[4],
    ci_upper = results$basic[5]
  )
}

if (subset_sample == TRUE) {
  trials_to_process <- all_exclusions_applied_ept_df |> slice_head(n = 500000)
} else {
  trials_to_process <- all_exclusions_applied_ept_df
}

EPT_scores <- trials_to_process |>
  mutate(
    session_id = as.factor(SESSION_ID),
    trial_consistency = as.character(trial_consistency)
  ) |>
  group_by(session_id, domain) |>
  filter(n_distinct(trial_consistency) == 2) |>
  do(bootstrap_EPT_score(data = .)) |>
  ungroup() |>
  mutate(
    sig = ifelse((ci_lower < 0 & ci_upper < 0) | (ci_lower > 0 & ci_upper > 0), TRUE, FALSE),
    ci_width = ci_upper - ci_lower
  )

```


# Process GNAT scoring data

```{r}

all_exclusions_applied_gnat_df <- read_rds("../../data/processed/data_trials_for_bootstrapping.rds") |>
  filter(measure %in% c("gnat"),
         BLOCK_NUMBER %in% c(3, 4, 7, 8)) 

```

## bootstrap GNAT scores

```{r}

n_boots <- 2000
subset_sample <- FALSE # optional for testing

if (TRUE) {
  
  # Function to compute RT score (inconsistent - consistent)
  GNAT_RT_score <- function(data, i) {
    data_with_indexes <- data[i, ]
    b <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "inconsistent"])
    a <- na.omit(data_with_indexes$score[data_with_indexes$trial_consistency == "consistent"])
    mean_a <- mean(a)
    mean_b <- mean(b)
    rt_score <- mean_b - mean_a
    return(rt_score)
  }

  bootstrap_GNAT_RT_score <- function(data) {
    require(dplyr)
    require(boot)
    
    fit <- boot::boot(
      data      = data,
      statistic = GNAT_RT_score,
      R         = n_boots,
      sim       = "ordinary",
      stype     = "i",
      parallel  = "multicore",
      ncpus     = parallel::detectCores() - 1
    )

    results <- boot::boot.ci(fit, conf = 0.95, type = "basic")
    
    tibble(
      method   = "basic",
      estimate = fit$t0,
      ci_lower = results$basic[4],
      ci_upper = results$basic[5]
    )
  }

  # Subset for speed (optional)
  if (subset_sample == TRUE) {
    trials_to_process <- all_exclusions_applied_gnat_df |>  # replace with your GNAT data frame
      slice_head(n = 500000)
  } else {
    trials_to_process <- all_exclusions_applied_gnat_df
  }

  # Bootstrap RT scores per participant + domain
  GNAT_RT_scores <- trials_to_process |>
    mutate(session_id = as.factor(SESSION_ID),
           trial_consistency = as.character(trial_consistency)) |>
    
    group_by(session_id, domain, trial_consistency) |>
    group_by(session_id, domain) |>
    filter(n_distinct(trial_consistency) == 2) |>  # Must have both conditions
    
    group_by(session_id, domain) |>
    do(bootstrap_GNAT_RT_score(data = .)) |>
    ungroup() |>
    mutate(
      sig = ifelse((ci_lower < 0 & ci_upper < 0) | (ci_lower > 0 & ci_upper > 0), TRUE, FALSE),
      ci_width = ci_upper - ci_lower
    )
}

```





# write to disk: IATs

```{r}

data_processed <- D_scores |>
  drop_na()
  
write_rds(data_processed, "../../data/processed/data_processed_iat_D.rds")

```

# write to disk: AMP

```{r}

write_rds(AMP_scores, "../../data/processed/data_processed_amp_native.rds")

```


# write to disk: EPT

```{r}

write_rds(EPT_scores, "../../data/processed/data_processed_ept_native.rds")

```

# write to disk: GNAT

```{r}

write_rds(GNAT_RT_scores, "../../data/processed/data_processed_gnat_native.rds")

```
